readme file completed
