--- code/src/artist.css
@@ -1,14 +1,0 @@
-.artist-name a {
-  color: #a0a0a0;
-  font-size: 14px;
-  font-family: Arial, Helvetica, sans-serif;
-  text-decoration: none;
-}
-
-.artist-name :hover {
-  text-decoration: underline;
-}
-
-.artist-box {
-  display: flex;
-}
