+++ code/src/components/Album.js
