+++ code/src/components/Header.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import './header.css';
 
 export const Header = () => {
   return (
