+++ code/src/components/MusicIcons.js
@@ -1,11 +1,10 @@
 import React from 'react';
-import { ReactComponent as HeartIcon } from './icons/heart.svg';
-import { ReactComponent as PlayIcon } from './icons/play.svg';
-import { ReactComponent as DotsIcon } from './icons/dots.svg';
+import { ReactComponent as HeartIcon } from '../icons/heart.svg';
+import { ReactComponent as PlayIcon } from '../icons/play.svg';
+import { ReactComponent as DotsIcon } from '../icons/dots.svg';
 
 export const MusicIcons = () => {
   return (
-    <>
       <section className="album-actions">
         <div className="album-actions-inner">
           <div className="album-actions-icon">
@@ -19,7 +18,6 @@ export const MusicIcons = () => {
           </div>
         </div>
       </section>
-    </>
   )
 };
